prepare JDBC branch

+++ src/main/java/com/endava/internship/dataCenter/controller/EmployeesController.java
@@ -16,9 +16,7 @@ public class EmployeesController {
 
     private final EmployeesService employeesService;
 
-
-
-//  GET(id)  -  get employee by id
+    //  GET(id)  -  get employee by id
     @GetMapping("/employees/{id}")
     public ResponseEntity<Employees> getEmployee(@PathVariable Integer id){
         if(id == null){
@@ -32,7 +30,6 @@ public class EmployeesController {
         }
 
         return new ResponseEntity<>(employee, HttpStatus.OK);
-        //return employeesService.getEmployeeJDBC(id);
     }
 
     // GET  -  list of all employees
@@ -47,11 +44,7 @@ public class EmployeesController {
         return new ResponseEntity<>(employeesList, HttpStatus.OK);
     }
 
-    ResponseEntity<String> firstAndLastNameNotNull() {
-        return new ResponseEntity<>("First name and Last name - not to be null, empty or blank", HttpStatus.BAD_REQUEST);
-    }
-
-//        First name and Last name - not to be null, empty or blank
+    //        First name and Last name - not to be null, empty or blank
 //        Email - to match email format
 //        Phone number - to start with 0 and and contain exactly 9 digits
 //        Salary - min 1.0
@@ -62,8 +55,11 @@ public class EmployeesController {
         //  First name and Last name - not to be null, empty or blank
         String checkFirstName = employeesDto.getFirstName();
         String checkLastName = employeesDto.getLastName();
-        if(checkFirstName == null && checkLastName == null && checkFirstName.trim().isEmpty() && checkLastName.trim().isEmpty()){
-            return new ResponseEntity<>("First name and Last name - not to be null, empty or blank", HttpStatus.BAD_REQUEST);
+        if(checkFirstName == null || checkLastName == null){
+            return new ResponseEntity<>("First name and Last name - not to be null", HttpStatus.BAD_REQUEST);
+        }
+        if(checkFirstName.trim().isEmpty() || checkLastName.trim().isEmpty()){
+            return new ResponseEntity<>("First name and Last name - empty or blank", HttpStatus.BAD_REQUEST);
         }
 
         // Email - to match email format
@@ -92,15 +88,58 @@ public class EmployeesController {
 
     // PUT - update an existing employee
     @PutMapping("/employees/{id}")
-    public Employees updateEmployeeById(@PathVariable Integer id, @RequestBody EmployeeDto employeeDto){
-        return employeesService.updateEmployees(id, employeeDto);
+    public ResponseEntity<String> updateEmployeeById(@PathVariable Integer id, @RequestBody EmployeeDto employeeDto){
+        //  First name and Last name - not to be null, empty or blank
+        String checkFirstName = employeeDto.getFirstName();
+        String checkLastName = employeeDto.getLastName();
+        if(checkFirstName == null || checkLastName == null){
+            return new ResponseEntity<>("First name and Last name - not to be null", HttpStatus.BAD_REQUEST);
+        }
+        if(checkFirstName.trim().isEmpty() || checkLastName.trim().isEmpty()){
+            return new ResponseEntity<>("First name and Last name - not to be empty or blank", HttpStatus.BAD_REQUEST);
+        }
+
+        // Email - to match email format
+        boolean checkEmail = employeeDto.getEmail().matches("[a-zA-Z0-9\\+\\.\\_\\%\\-\\+]{1,256}" + "\\@" + "[a-zA-Z0-9][a-zA-Z0-9\\-]{0,64}" + "("
+                + "\\." + "[a-zA-Z0-9][a-zA-Z0-9\\-]{0,25}" + ")+");
+        if(!checkEmail){
+            return new ResponseEntity<>("Email is not correct", HttpStatus.BAD_REQUEST);
+        }
+
+        // Phone number - to start with 0 and and contain exactly 9 digits
+        boolean checkPhoneNumber = employeeDto.getPhoneNumber().matches("^0\\d{9}");
+        if(!checkPhoneNumber){
+            return new ResponseEntity<>("Phone number is not correct. Need to start with 0 and contain exactly 9 digits",
+                    HttpStatus.BAD_REQUEST);
+        }
+
+        // Salary - min 1.0
+        boolean checkSalary = employeeDto.getSalary() >= 1.0;
+        if(!checkSalary){
+            return new ResponseEntity<>("Check salary. Min value is 1.0", HttpStatus.BAD_REQUEST);
+        }
+
+        employeesService.updateEmployees(id, employeeDto);
+
+        return new ResponseEntity<>("Employee updated", HttpStatus.OK);
     }
 
 
     // DELETE  -  delete employee by id
     @DeleteMapping("/employees/{id}")
-    public void deleteEmployee(@PathVariable Integer id){
+    public ResponseEntity<String> deleteEmployee(@PathVariable Integer id){
+        if(id == null){
+            return new ResponseEntity<>(HttpStatus.BAD_REQUEST);
+        }
         employeesService.deleteEmployee(id);
+
+        Employees employee = this.employeesService.getEmployeeById(id);
+
+        if(employee == null){
+            return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+        }
+
+        return new ResponseEntity<>("Employee deleted", HttpStatus.NO_CONTENT);
     }
 
 }
